refactor(material): extract SearchView.svelte from MaterialPopup
Move the search input + results list + custom-material edit/delete
handlers out of MaterialPopup. The scoring, material list derivations,
and select/useQuery helpers move with it.

The InspectPanel row between input and list is preserved via an optional
betweenInputAndResults snippet slot. The parent still owns query (passed
in + onQueryChange) and opens the edit flow via an oneditRequest callback
that populates the in-parent DefineForm state.

Parent exposes a focus() function on SearchView via bind:this; the
auto-focus effect calls it verbatim. No behavior change.

Refs: #64 Phase 0
